requirements.txt
streamlit
pdfplumber
langchain
langchain-community
chromadb
sentence-transformers
google-generativeai
